# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/faq": {
            "get": {
                "tags": [
                    "Faq"
                ],
                "summary": "Get list of Faq",
                "description": "Returns list of Faq",
                "operationId": "Faq",
                "responses": {
                    "201": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Faq"
                ],
                "summary": "Store Faq",
                "description": "Returns Faq data",
                "operationId": "FaqStore",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {}
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Faq created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Faq created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/faq/{id}": {
            "get": {
                "tags": [
                    "Faq"
                ],
                "summary": "Get Faq information",
                "description": "Returns Faq information",
                "operationId": "FaqGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Faq id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Faq"
                ],
                "summary": "Update existing Faq",
                "description": "Returns updated Faq data",
                "operationId": "FaqUdpate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Faq updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Faq updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Faq"
                ],
                "summary": "Delete existing Faq",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteFaq",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Faq id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Faq deleted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http,https",
                "name": "Authorization",
                "bearerFormat": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": {
        "passport": {
            "type": "apiKey",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "http,https",
            "name": "Authorization",
            "bearerFormat": "bearer",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}